Implement post and read functinality

+++ Board.cpp
@@ -82,7 +82,7 @@ std::string Board::read(unsigned int row, unsigned int column, Direction d , uns
 
 void Board::show(){
 
-	for(int i=0;i<board->size(); ++i){
+	for(size_t i=0;i<board->size(); ++i){
 		std::cout << board->at(i) << endl;
 	}
 }
